[1201] FIX 1. Fixed HandleModDamagePercentDone function

+++ src/game/SpellAuras.cpp
@@ -1248,18 +1248,18 @@ void Aura::HandleModDamagePercentDone(bool apply)
 {
     if(m_modifier->m_miscvalue == 1)
     {
-        m_target->SetUInt32Value(UNIT_FIELD_MINDAMAGE, (uint32)(m_target->GetUInt32Value(UNIT_FIELD_MINDAMAGE) * (apply ? (100.0f+m_modifier->m_amount)/100.0f : 100.0f / (100.0f+m_modifier->m_amount))) );
-        m_target->SetUInt32Value(UNIT_FIELD_MAXDAMAGE, (uint32)(m_target->GetUInt32Value(UNIT_FIELD_MAXDAMAGE) * (apply ? (100.0f+m_modifier->m_amount)/100.0f : 100.0f / (100.0f+m_modifier->m_amount))) );
+        m_target->SetFloatValue(UNIT_FIELD_MINDAMAGE, (uint32)(m_target->GetFloatValue(UNIT_FIELD_MINDAMAGE) * (apply ? (100.0f+m_modifier->m_amount)/100.0f : 100.0f / (100.0f+m_modifier->m_amount))) );
+		m_target->SetFloatValue(UNIT_FIELD_MAXDAMAGE, (uint32)(m_target->GetFloatValue(UNIT_FIELD_MAXDAMAGE) * (apply ? (100.0f+m_modifier->m_amount)/100.0f : 100.0f / (100.0f+m_modifier->m_amount))) );
     }
     if(m_modifier->m_miscvalue == 126)
     {
-        m_target->SetUInt32Value(UNIT_FIELD_MINOFFHANDDAMAGE, (uint32)(m_target->GetUInt32Value(UNIT_FIELD_MINOFFHANDDAMAGE) * (apply?(100.0f+m_modifier->m_amount)/100.0f : 100.0f / (100.0f+m_modifier->m_amount))) );
-        m_target->SetUInt32Value(UNIT_FIELD_MAXOFFHANDDAMAGE, (uint32)(m_target->GetUInt32Value(UNIT_FIELD_MAXOFFHANDDAMAGE) * (apply?(100.0f+m_modifier->m_amount)/100.0f : 100.0f / (100.0f+m_modifier->m_amount))) );
+        m_target->SetFloatValue(UNIT_FIELD_MINOFFHANDDAMAGE, (uint32)(m_target->GetFloatValue(UNIT_FIELD_MINOFFHANDDAMAGE) * (apply?(100.0f+m_modifier->m_amount)/100.0f : 100.0f / (100.0f+m_modifier->m_amount))) );
+        m_target->SetFloatValue(UNIT_FIELD_MAXOFFHANDDAMAGE, (uint32)(m_target->GetFloatValue(UNIT_FIELD_MAXOFFHANDDAMAGE) * (apply?(100.0f+m_modifier->m_amount)/100.0f : 100.0f / (100.0f+m_modifier->m_amount))) );
     }
     if(m_modifier->m_miscvalue == 127)
     {
-        m_target->SetUInt32Value(UNIT_FIELD_MINRANGEDDAMAGE, (uint32)(m_target->GetUInt32Value(UNIT_FIELD_MINRANGEDDAMAGE) * (apply?(100.0f+m_modifier->m_amount)/100.0f : 100.0f / (100.0f+m_modifier->m_amount))) );
-        m_target->SetUInt32Value(UNIT_FIELD_MAXRANGEDDAMAGE, (uint32)(m_target->GetUInt32Value(UNIT_FIELD_MAXRANGEDDAMAGE) * (apply?(100.0f+m_modifier->m_amount)/100.0f : 100.0f / (100.0f+m_modifier->m_amount))) );
+        m_target->SetFloatValue(UNIT_FIELD_MINRANGEDDAMAGE, (uint32)(m_target->GetFloatValue(UNIT_FIELD_MINRANGEDDAMAGE) * (apply?(100.0f+m_modifier->m_amount)/100.0f : 100.0f / (100.0f+m_modifier->m_amount))) );
+        m_target->SetFloatValue(UNIT_FIELD_MAXRANGEDDAMAGE, (uint32)(m_target->GetFloatValue(UNIT_FIELD_MAXRANGEDDAMAGE) * (apply?(100.0f+m_modifier->m_amount)/100.0f : 100.0f / (100.0f+m_modifier->m_amount))) );
     }
 }
